prevent to draft invoice if still have credit notes

+++ netforce_account/netforce_account/models/account_invoice.py
@@ -523,10 +523,10 @@ class Invoice(Model):
 
     def to_draft(self, ids, context={}):
         obj = self.browse(ids)[0]
-        for credit_note in obj.credit_notes:
-            credit_note.delete()
         if obj.state != "waiting_payment":
             raise Exception("Invalid status")
+        if obj.credit_notes:
+            raise Exception("There are still payment entries for this invoice")
         if obj.move_id:
             obj.move_id.void()
             obj.move_id.delete()
